# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GUG Apis ",
        "contact": {
            "name": "Yosof Bayan",
            "url": "https://wa.me/+963967213544",
            "email": "yosofbayan75@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "local Base URL"
        },
        {
            "url": "http://gug-back.icrcompany.net/api/v1",
            "description": "Develop Base URL"
        }
    ],
    "paths": {
        "/admin/about-us": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - AboutUs"
                ],
                "summary": "get all about-us",
                "operationId": "3d4d14d08ac47da841b7a7c4a7f8c62b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AboutUsResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - AboutUs"
                ],
                "summary": "Create a new AboutUs",
                "operationId": "679b0ee11201bc3c8ab7a396e5c4187c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAboutUsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AboutUsResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/about-us/{about}": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - AboutUs"
                ],
                "summary": "show an about",
                "operationId": "d55d7b90ce902aa379a748e042795b9c",
                "parameters": [
                    {
                        "name": "about",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AboutUsResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - AboutUs"
                ],
                "summary": "update an existing  AboutUs",
                "operationId": "5b89d57b9d1cfda4bdd50fe8304e9b0d",
                "parameters": [
                    {
                        "name": "about",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT",
                            "enum": [
                                "PUT"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAboutUsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AboutUsResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/about-us/{id}": {
            "delete": {
                "tags": [
                    "Admin",
                    "Admin - AboutUs"
                ],
                "summary": "SoftDelete an existing AboutUs",
                "operationId": "d32bfb504e38a1ff911d38d75707c3ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateAboutUsRequest": {
                "required": [
                    "title",
                    "icon",
                    "background_color"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "background_color": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "description": {},
                    "icon": {
                        "type": "string",
                        "format": "binary",
                        "example": "image.jpg"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "example": "image.jpg"
                    },
                    "offices[0][address]": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "offices[0][phone_number]": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    }
                },
                "type": "object"
            },
            "UpdateAboutUsRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "background_color": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "description": {},
                    "icon": {
                        "type": "string",
                        "format": "binary",
                        "example": "image.jpg"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "example": "image.jpg"
                    },
                    "offices[0][address]": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "offices[0][phone_number]": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "delete_offices[0]": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "AboutUsResource": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "backgroun_color": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "description": {},
                    "icon": {
                        "type": "string",
                        "format": "binary",
                        "example": "image.jpg"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "example": "image.jpg"
                    },
                    "offices[0][address]": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    },
                    "offices[0][phone_number]": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sample Name"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}